+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/right-panel/types.tsx
@@ -1,0 +1,8 @@
+import { ComponentType } from "react";
+export type FormItem = {
+  id: string;
+  name: string;
+  icon: ComponentType | null;
+  readOnly: boolean;
+  children?: FormItem[];
+};
